+++ src/main/kotlin/us/frollo/frollosdk/database/SDKDatabase.kt
@@ -102,7 +102,7 @@ import us.frollo.frollosdk.model.coredata.user.User
         Address::class,
         ServiceOutage::class
     ],
-    version = 16, exportSchema = true
+    version = 17, exportSchema = true
 )
 
 @TypeConverters(Converters::class)
@@ -174,7 +174,8 @@ abstract class SDKDatabase : RoomDatabase() {
                     MIGRATION_12_13,
                     MIGRATION_13_14,
                     MIGRATION_14_15,
-                    MIGRATION_15_16
+                    MIGRATION_15_16,
+                    MIGRATION_16_17
                 )
                 .build()
         }
@@ -675,6 +676,16 @@ abstract class SDKDatabase : RoomDatabase() {
                 database.execSQL("CREATE INDEX IF NOT EXISTS `index_addresses_address_id` ON `addresses` (`address_id`)")
                 database.execSQL("COMMIT")
                 // END - Drop & re-create table addresses
+            }
+        }
+
+        private val MIGRATION_16_17: Migration = object : Migration(16, 17) {
+            override fun migrate(database: SupportSQLiteDatabase) {
+
+                // New changes in this migration:
+                // 1) Alter table - user - add middle_names column
+
+                database.execSQL("ALTER TABLE `user` ADD COLUMN `middle_names` TEXT")
             }
         }
     }